# BGC_Portal/BCG_WebApi_swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BGC_WebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/RpgSession": {
      "get": {
        "tags": [
          "RpgSession"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RpgSessionDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RpgSessionDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RpgSessionDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RpgSession"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RpgSessionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RpgSessionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RpgSessionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "RpgSession"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RpgSessionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RpgSessionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RpgSessionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "RpgSession"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/RpgSession/{id}": {
      "get": {
        "tags": [
          "RpgSession"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RpgSessionDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RpgSessionDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RpgSessionDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Game": {
        "type": "object",
        "properties": {
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "minNumberPlayers": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxNumberPlayers": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameVersion"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GameSession": {
        "type": "object",
        "properties": {
          "gameSessionId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "schedule": {
            "type": "string",
            "format": "date-time"
          },
          "minNumberPlayers": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxNumberPlayers": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "length": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "organizerId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "organizer": {
            "$ref": "#/components/schemas/Member"
          },
          "gameTableId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "gameTable": {
            "$ref": "#/components/schemas/GameTable"
          },
          "gameSessionRegistrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameSessionRegistration"
            },
            "nullable": true
          },
          "gamesSessionGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameSessionGame"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GameSessionGame": {
        "type": "object",
        "properties": {
          "gameSessionGameId": {
            "type": "integer",
            "format": "int32"
          },
          "gameSessionId": {
            "type": "integer",
            "format": "int32"
          },
          "gameSession": {
            "$ref": "#/components/schemas/GameSession"
          },
          "gameVersionId": {
            "type": "integer",
            "format": "int32"
          },
          "gameVersion": {
            "$ref": "#/components/schemas/GameVersion"
          }
        },
        "additionalProperties": false
      },
      "GameSessionRegistration": {
        "type": "object",
        "properties": {
          "gameSessionRegistrationId": {
            "type": "integer",
            "format": "int32"
          },
          "memberId": {
            "type": "integer",
            "format": "int32"
          },
          "player": {
            "$ref": "#/components/schemas/Member"
          },
          "gameSessionId": {
            "type": "integer",
            "format": "int32"
          },
          "gameSession": {
            "$ref": "#/components/schemas/GameSession"
          }
        },
        "additionalProperties": false
      },
...
      "GameTable": {
        "type": "object",
        "properties": {
          "gameTableId": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "gameSessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameSession"
            },
            "nullable": true
          },
          "rpgSessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RpgSession"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "GameVersion": {
        "type": "object",
        "properties": {
          "gameVersionId": {
            "type": "integer",
            "format": "int32"
          },
          "gameId": {
            "type": "integer",
            "format": "int32"
          },
          "game": {
            "$ref": "#/components/schemas/Game"
          },
          "gameSessionGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameSessionGame"
            },
            "nullable": true
          },
          "rpgSessionGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RpgSessionGame"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Member": {
        "type": "object",
        "properties": {
          "memberId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "nullable": true
          },
          "rpgCampaigns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RpgCampaign"
            },
            "nullable": true
          },
          "gameSessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameSession"
            },
            "nullable": true
          },
          "rpgCampaignRegistrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RpgCampaignRegistration"
            },
            "nullable": true
          },
          "gameSessionRegistrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameSessionRegistration"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RpgCampaign": {
        "type": "object",
        "properties": {
          "rpgCampaignId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "organizerId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "organizer": {
            "$ref": "#/components/schemas/Member"
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RpgSession"
            },
            "nullable": true
          },
          "rpgCampaignRegistrations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RpgCampaignRegistration"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RpgCampaignRegistration": {
        "type": "object",
        "properties": {
          "rpgCampaignRegistrationId": {
            "type": "integer",
            "format": "int32"
          },
          "memberId": {
            "type": "integer",
            "format": "int32"
          },
          "player": {
            "$ref": "#/components/schemas/Member"
          },
          "rpgCampaignId": {
            "type": "integer",
            "format": "int32"
          },
          "rpgCampaign": {
            "$ref": "#/components/schemas/RpgCampaign"
          }
        },
        "additionalProperties": false
      },
...
      "RpgSession": {
        "type": "object",
        "properties": {
          "rpgSessionId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "schedule": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "rpgCampaignId": {
            "type": "integer",
            "format": "int32"
          },
          "campaign": {
            "$ref": "#/components/schemas/RpgCampaign"
          },
          "gameTableId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "table": {
            "$ref": "#/components/schemas/GameTable"
          },
          "rpgSessionGames": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RpgSessionGame"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RpgSessionDTO": {
        "type": "object",
        "properties": {
          "rpgSessionId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "schedule": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "rpgCampaignId": {
            "type": "integer",
            "format": "int32"
          },
          "campaign": {
            "$ref": "#/components/schemas/RpgCampaign"
          },
          "gameTableId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "table": {
            "$ref": "#/components/schemas/GameTable"
          }
        },
        "additionalProperties": false
      },
      "RpgSessionGame": {
        "type": "object",
        "properties": {
          "rpgSessionGameId": {
            "type": "integer",
            "format": "int32"
          },
          "rpgSessionId": {
            "type": "integer",
            "format": "int32"
          },
          "rpgSession": {
            "$ref": "#/components/schemas/RpgSession"
          },
          "gameVersionId": {
            "type": "integer",
            "format": "int32"
          },
          "gameVersion": {
            "$ref": "#/components/schemas/GameVersion"
          }
        },
        "additionalProperties": false
      },
      "SessionStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      }
    }
  }
}